requirements.txt
# Cơ bản & Xử lý ảnh
numpy==1.26.4
Pillow
opencv-python
shapely
pyclipper
tqdm

# Deep Learning Frameworks
torch
torchvision
torchaudio
torchsummary

# OCR & Inpainting
manga-ocr
simple-lama-inpainting
ultralytics

# AI & API
google-generativeai
pydantic
pydantic-settings
python-dotenv

# NLP
sudachipy
sudachidict_full
pykakasi

# Tiện ích
wandb
requests

# Test
pytest